+++ pkg/service/manifest/model/model.go
@@ -22,7 +22,7 @@ type CreateManifestRequest struct {
 	IssuerName        *string                        `json:"issuerName,omitempty"`
 	OutputDescriptors []manifestsdk.OutputDescriptor `json:"outputDescriptors" validate:"required,dive"`
 	ClaimFormat       *exchange.ClaimFormat          `json:"format" validate:"required,dive"`
-	// TODO: Allow for specifying the presentation definition only by id.
+	// TODO: Allow for specifying the presentation definition only by id. https://github.com/TBD54566975/ssi-service/issues/491
 	PresentationDefinition *exchange.PresentationDefinition `json:"presentationDefinition,omitempty" validate:"omitempty,dive"`
 }
 
@@ -164,8 +164,6 @@ type DeleteRequestRequest struct {
 	ID string `json:"id" validate:"required"`
 }
 
-type ListRequestsRequest struct{}
-
 type ListRequestsResponse struct {
 	ManifestRequests []Request `json:"manifestRequests"`
 }
@@ -179,5 +177,5 @@ type Request struct {
 	// CredentialManifestJWT is a JWT token with a "presentation_definition" claim within it. The
 	// value of the field named "presentation_definition.id" matches PresentationDefinitionID.
 	// This is an output only field.
-	CredentialManifestJWT keyaccess.JWT `json:"CredentialManifestJwt"`
+	CredentialManifestJWT keyaccess.JWT `json:"credentialManifestJwt"`
 }
